+++ aptos-move/aptos-vm/src/data_cache.rs
@@ -4,7 +4,7 @@
 //! Scratchpad for on chain values during the execution.
 
 use crate::{
-    aptos_vm_impl::gas_config,
+    aptos_vm::gas_config,
     move_vm_ext::{get_max_binary_format_version, MoveResolverExt},
 };
 #[allow(unused_imports)]
